+++ boardify-service/src/main/java/com/boardify/boardify_service/common/event/ListUpdatedEvent.java
@@ -1,3 +1,3 @@
 package com.boardify.boardify_service.common.event;
 
-public class ListUpdatedEvent { public Long listId; public String name; public int position; public String version = "1"; }
+public class ListUpdatedEvent { public Long listId; public String name; public double position; public String version = "1"; }
