+++ boardify-service/src/main/java/com/boardify/boardify_service/list/dto/ListDto.java
@@ -1,4 +1,4 @@
 package com.boardify.boardify_service.list.dto;
 
-public record ListDto(Long id, Long boardId, String name, int position) {}
+public record ListDto(Long id, Long boardId, String name, double position) {}
